#279,#280 - Add ticket and visa concept
- session and authentication on services

+++ Tourine.ServiceInterfaces/Currency.cs
@@ -7,6 +7,6 @@ namespace Tourine.ServiceInterfaces
         [AutoIncrement]
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Factor { get; set; }
+        public float Factor { get; set; }
     }
 }
